requirements.txt
faiss-cpu
numpy
pyyaml
sqlite-utils
ollama
rank-bm25
